requirements.txt
sounddevice
numpy
scipy
python-dotenv
numba
librosa
colorama
elevenlabs
keyboard
pydub